requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.1
Django==6.0.1
django-cloudinary-storage==0.3.0
django-taggit==6.1.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
waitress==3.0.2
whitenoise==6.11.0
